Added Clifford's changes involving Polyglot loading and error reporting. Not building yet.
Polyglot gem wouldn't install at time of patch so I am not going to put it as a dependency
in the gem yet, but it seems like that can be resolved. Maybe there's a rubyforge problem.
The error reporting needs to be modernized based on the optimization removing nested failures.
This can be merged in when that is dealt with.