Very inefficient correction to NodeCache to correctly expire non-terminal results.
Before I was registering only terminal nodes for detection of interval disturbance, but I forgot to account for the fact that a result can be invalidated by disturbing the buffer BETWEEN two of its child terminal nodes, an act that would not be detected as a disturbance in the old scheme. All nodes need to be checked for disturbance... I cannot only check terminals and chain expiration to their parents. Non-local dependencies are also still possible, so the chaining scheme must also remain in place. I have not sorted out what will be expired of its own right and what needs to use expiration propagation via the dependencies mechanism, so the existing solution allows multiple expirations of a node, on the basis of its interval being disturbed AND on the basis of its dependencies being expired. This is just to get the behavior working. It needs to be tightened up.