Sequences ending in an epsilon result have intervals that include their endpoint.
Because an epsilon indicates reliance on not a range, but a point in the buffer, it is important that the disturbance of this point cause the expiration of a result that depends upon the epsilon node. Whereas normal results have a range, which would naturally be contained by the range of the parent result, epsilons can only be contained by the range of the parent if they include their endpoint.